Add files via upload
To use this program, create a folder ("mds") and add the above files (after decompressing). The compressed folder "sounds.zip" should be put within the "resources.zip" folder. Github was unable to upload these folders together due to size limitations. 

The python program is meant to be run out of Terminal by calling the folder "mds."